clean up
cleaned up require relative. Moved term milestones to a new file, just because of the ascii art.
