+++ end_game.rb
@@ -1,4 +1,3 @@
-
 def end_of_game
   system 'clear'
   puts "This is the end of the game.... in end_game.rb"
